Update testcode
This code is part of the Options Pricing Project, developed by Purdue Momentum in Spring '25. The goal of the project was to backtest a trading strategy using the pricing tool; comparing theoretical option values with market values, calculating technical indicators, and more. While the team wasn't able to fully develop the trading strategy, this responsibility has been handed off to the new members of Purdue Momentum, who now have a solid foundation to continue the work.

The code retrieves stock price data through APIs and calculates theoretical option prices using the Black-Scholes model for European options and binomial trees for American options. It leverages historical volatility from over 500 trading days, which is an area identified for improvement (there's a need for a better volatility measure). Additionally, the code computes the Greeks (Delta, Gamma, Vega, Rho, Theta). This project provides a strong base for future members to complete and refine.